Added gitlab, tutorial downloader can now accept more sources
